fix(painter-dom): break list marker renderer cycle
